+++ VisualStudio/tools/extractor/sources.props
@@ -1,6 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <ItemGroup>
+    <ClCompile Include="..\engine\agg_file.cpp" />
     <ClCompile Include="..\engine\logging.cpp" />
     <ClCompile Include="..\engine\serialize.cpp" />
     <ClCompile Include="..\engine\system.cpp" />
@@ -9,6 +10,7 @@
     <ClCompile Include="extractor.cpp" />
   </ItemGroup>
   <ItemGroup>
+    <ClInclude Include="..\engine\agg_file.h" />
     <ClInclude Include="..\engine\endian_h2.h" />
     <ClInclude Include="..\engine\logging.h" />
     <ClInclude Include="..\engine\math_base.h" />
