+++ src/engine/agg_file.h
@@ -26,6 +26,7 @@
 #include <functional>
 #include <map>
 #include <string>
+#include <string_view>
 #include <utility>
 #include <vector>
 
@@ -69,6 +70,8 @@ namespace fheroes2
         uint8_t animationFrames; // used for adventure map animations, this can replace ICN::AnimationFrame
         uint32_t offsetData;
     };
+
+    uint32_t calculateAggFilenameHash( const std::string_view str );
 }
 
 StreamBase & operator>>( StreamBase & st, fheroes2::ICNHeader & icn );
